requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.14.0
openpyxl>=3.1.0
xlrd>=2.0.0
yfinance
sgs-data>=2.0.0